requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
requests==2.22.0
Flask-Testing==0.7.1